Implement public key export
